+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * github-204: [BUG] Detect indirect dependencies in a chain of Hive views
 * github-207: [BUG] Build should not directly fail if infering dirty status fails
 * github-209: [BUG] HiveViews should not trigger cascaded refresh during CREATE phase even when nothing is changed
+* github-211: Implement new hiveQuery relation
 
 ## Breaking changes